+++ src/Schema/Type/BlockType.php
@@ -55,6 +55,11 @@ final class BlockType extends AbstractType implements DatabaseTypeInterface
     /**
      * @var string
      */
+    private $template;
+
+    /**
+     * @var string
+     */
     private $blockType;
 
     public function __construct(
@@ -153,7 +158,11 @@ final class BlockType extends AbstractType implements DatabaseTypeInterface
     public function __toString()
     {
         try {
-            return $this->renderer->render($this->getBlock()->template(), $this->value());
+            $template = $this->getBlock()->template();
+            if ($this->template) {
+                $template = $this->template;
+            }
+            return $this->renderer->render($template, $this->value());
         } catch (\Throwable $e) {
             if (!$this->applicationConfig->isDevelopment()) {
                 return '';
@@ -188,6 +197,11 @@ final class BlockType extends AbstractType implements DatabaseTypeInterface
         return JsonType::class;
     }
 
+    public function type(): string
+    {
+        return $this->getBlock()->serviceName();
+    }
+
     public function jsonSerialize()
     {
         return \array_merge(
@@ -210,6 +224,14 @@ final class BlockType extends AbstractType implements DatabaseTypeInterface
             'value' => $this->value,
             'blockType' => $this->blockType,
         ]);
+    }
+
+    public function withTemplate(string $template): BlockType
+    {
+        $block = clone $this;
+        $block->template = $template;
+
+        return $block;
     }
 
     public function unserialize($serialized)