+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 
 ## [Unreleased](https://github.com/ixocreate/cms-package/compare/0.3.1...develop)
+### Added
+- change Template in Block
 
 ## [v0.3.1 (2019-05-21)](https://github.com/ixocreate/cms-package/compare/0.3.0...0.3.1)
 ### Added